// src/api/openapi/model/offerStatementEntryEntry.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface OfferStatementEntryEntry { 
    additionSubstraction: boolean;
    amount: number;
    description: string;
    id: string;
    price: number;
}